requirements.txt
tensorflow-gpu==2.10.1
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.19.0
transformers==4.30.0 